requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
openai-whisper==20231117
torch==2.2.1
pytest==7.4.4
httpx==0.27.0
numpy==1.24.3